Move and rename connectionOptions to utils/generateConnectionOptions
